refactor: rename /src/validation to /src/core and remove tr- prefix from core files
- Rename directory /src/validation to /src/core
- Rename tr-validation.ts to validate.ts
- Rename tr-input.ts to input.ts
- Rename tr-bag.ts to bag.ts
- Rename tr-form.ts to form.ts
- Update all imports and exports across the project
- Update test files to match new naming convention
- Update readme.md documentation

Fixes #13
