+++ src/core/utils/rule-executed.ts
