+++ src/core/validate.ts
