+++ src/messages/index.ts
@@ -1,2 +1,2 @@
 export * from './tr-messages';
-export * from '../validation/utils/input-rule';
+export * from '../core/utils/input-rule';
