--- src/validation/index.ts
@@ -1,8 +1,0 @@
-export * from './abstract-input';
-export * from './tr-input';
-export * from './tr-form';
-export * from './trivule';
-export * from './tr-validation';
-export * from './tr-bag';
-
-export * from './utils/rule-executed';
