+++ tests/input.test.ts
@@ -1,4 +1,4 @@
-import { TrivuleInput } from '../src/validation/tr-input';
+import { TrivuleInput } from '../src/core/input';
 
 describe('TrivuleInput', () => {
   describe('getRules', () => {
